Refactor SWAT science metadata into a separate test case
